refactor + finish weather
